old_gemini/requirements.txt
google-genai>=1.0.0
fastmcp>=2.0.0
